유기동물 메인에 정보가 깔끔하게 나오게 업데이트
유기동물 보호소, 품종까지 깔끔하게 출력!

+++ src/main/java/sist/com/vo/AbShelterVO.java
@@ -16,11 +16,11 @@ public class AbShelterVO {
 	private String abShelterName;
 	private int sidoCode;
 	private int sigunguCode;
-	private int abShelterTel;
+	private long abShelterTel;
 	private String abShelterAddress;
 	private String organize;
 	private String managerName;
-	private int managerTel;
+	private long managerTel;
 }                                     
 /*
